[이영호] 회원삭제 패스워드 확인절차 추가
1.DeleteMemberController.java, deleteMember 회원삭제 패스워드 확인절차 코드추가
2. DBHelper.java mybatis메뉴얼에 따른 코드 변경(SqlSessionFactory 여러번 빌드하지 않게수정
http://www.mybatis.org/mybatis-3/ko/getting-started.html 에서 [스코프 와 생명주기
참고)
3. OrderListController.java, MemberItemDao.java, MemberItemMapper.xml
관리자는 모든 주문목록을 볼 수 있고, 고객은 자신이 주문한 목록만 볼 수 있게 수정

+++ mymall/src/com/test/mymall/web/DeleteMemberController.java
@@ -14,14 +14,31 @@ import com.test.mymall.vo.Member;
 public class DeleteMemberController extends HttpServlet {
 	private MemberService memberService;
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+		System.out.println("DeleteMemberController.doGet()");
 		HttpSession session = request.getSession();
 		if(session.getAttribute("loginMember") != null) {
-			memberService = new MemberService();
-			Member member = (Member)session.getAttribute("loginMember");
-			memberService.removeMember(member.getNo());
-			session.invalidate();
+			request.getRequestDispatcher("/WEB-INF/views/deleteMember.jsp").forward(request, response);;
+		}
+		else {
 			response.sendRedirect(request.getContextPath() + "/IndexController");
 		}
+	}
+	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+		System.out.println("DeleteMemberController.doPost()");
+		HttpSession session = request.getSession();
+		if(session.getAttribute("loginMember") != null) {
+			String pw = request.getParameter("pw");
+			memberService = new MemberService();
+			Member member = (Member)session.getAttribute("loginMember");
+			member.setPw(pw);
+			member = memberService.loginMember(member);
+			if(member != null) {
+				memberService.removeMember(member.getNo());
+				session.invalidate();
+			}
+
+		}
+		response.sendRedirect(request.getContextPath() + "/IndexController");
 
 	}
 }
